+++ Sources/TraceLogObjC/TraceLog.h
@@ -126,7 +126,7 @@
 #if !TRACELOG_DISABLED
     #define LogIfEnabled(logLevel,tagName,format,...) [TLLogger logPrimitive: logLevel tag: tagName file: [NSString stringWithUTF8String: __FILE__] function: [NSString stringWithUTF8String: __FUNCTION__] line: __LINE__ message: ^{ return [NSString stringWithFormat: format, ##__VA_ARGS__]; }]
 #else
-    #define LogIfEnabled(logLevel,label, format, ...) ((void)0)
+    #define LogIfEnabled(logLevel,tagName,format, ...) ((void)0)
 #endif
 
 /// Low level - for use in mixed low level C code.