+++ Tests/LinuxMain.swift
@@ -17,8 +17,6 @@ extension TraceLogPerformanceTestsSwift {
 
    static var allTests: [(String, (TraceLogPerformanceTestsSwift) -> () throws -> Void)] {
       return [
-                ("testLogErrorPerformance", testLogErrorPerformance),
-                ("testLogTrace4Performance", testLogTrace4Performance),
                 ("testLogErrorPerformance_NullWriter", testLogErrorPerformance_NullWriter),
                 ("testLogTrace4Performance_NullWriter", testLogTrace4Performance_NullWriter)
            ]